pyproject.toml
[project]
name = "minichess"
version = "0.1.0"
requires-python = ">=3.11"
description = "Floating desktop chess app"
dependencies = [
    "PySide6>=6.7.0",
    "python-chess>=1.999",
    "pystray>=0.19.5",
    "Pillow>=10.0.0",
    "requests>=2.31.0",
]

[project.scripts]
minichess = "minichess.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/minichess"]
